move unit tests into __tests__ directory
